refactor(account-keys): share quick-list presentation policy
