+++ src/features/AccountManagement/components/CopyKeyDialog/RuntimeKeyItem.tsx
@@ -7,15 +7,19 @@ import { useTranslation } from "react-i18next"
 
 import { Badge, Card, CardContent, IconButton } from "~/components/ui"
 import { getCopyKeyDialogRuntimeKeyItemTestId } from "~/features/AccountManagement/testIds"
+import { KeyResourceCard } from "~/features/KeyManagement/components/KeyResourceCard"
+import { buildLegacyKeyResourceCardPresentation } from "~/features/KeyManagement/presentation/legacyKeyResourceCard"
 import {
   ACCOUNT_RUNTIME_KEY_STATUSES,
   isAccountTokenRuntimeKey,
   type AccountRuntimeKey,
+  type AccountTokenRuntimeKey,
 } from "~/services/accounts/accountRuntimeKeys"
 import type { ApiToken, DisplaySiteData } from "~/types"
 import { getGroupBadgeStyle } from "~/utils/core/formatters"
 
-import { RuntimeKeyDetails } from "./RuntimeKeyDetails"
+import { RuntimeKeyActionControls } from "./RuntimeKeyActionControls"
+import { RuntimeKeyDetails, RuntimeKeySecretPreview } from "./RuntimeKeyDetails"
 
 const getRuntimeKeyStatusBadgeStyle = (
   runtimeKey: Pick<AccountRuntimeKey, "status">,
@@ -47,9 +51,22 @@ export function RuntimeKeyItem({
   onOpenCCSwitchDialog,
 }: RuntimeKeyItemProps) {
   const { t } = useTranslation("ui")
-  const group = isAccountTokenRuntimeKey(runtimeKey)
-    ? runtimeKey.token.group
-    : ""
+
+  if (isAccountTokenRuntimeKey(runtimeKey)) {
+    return (
+      <AccountTokenRuntimeKeyItem
+        runtimeKey={runtimeKey}
+        isExpanded={isExpanded}
+        copiedRuntimeKeyId={copiedRuntimeKeyId}
+        onToggle={onToggle}
+        onCopyKey={onCopyKey}
+        account={account}
+        onOpenCCSwitchDialog={onOpenCCSwitchDialog}
+      />
+    )
+  }
+
+  const group = ""
   const isActive = runtimeKey.status === ACCOUNT_RUNTIME_KEY_STATUSES.Active
 
   return (
@@ -115,5 +132,48 @@ export function RuntimeKeyItem({
         />
       )}
     </Card>
+  )
+}
+
+/** Composes a legacy account runtime key through the shared key-resource card. */
+function AccountTokenRuntimeKeyItem({
+  runtimeKey,
+  isExpanded,
+  copiedRuntimeKeyId,
+  onToggle,
+  onCopyKey,
+  account,
+  onOpenCCSwitchDialog,
+}: Omit<RuntimeKeyItemProps, "runtimeKey"> & {
+  runtimeKey: AccountTokenRuntimeKey
+}) {
+  const { t } = useTranslation("keyManagement")
+  const presentation = buildLegacyKeyResourceCardPresentation(runtimeKey, t)
+
+  return (
+    <KeyResourceCard
+      presentation={presentation}
+      secret={
+        presentation.actions.copySecret ? (
+          <RuntimeKeySecretPreview secret={runtimeKey.secret} />
+        ) : presentation.maskedLabel ? (
+          <code>{presentation.maskedLabel}</code>
+        ) : undefined
+      }
+      secretControls={
+        <RuntimeKeyActionControls
+          runtimeKey={runtimeKey}
+          actionPolicy={presentation.actions}
+          copiedRuntimeKeyId={copiedRuntimeKeyId}
+          onCopyKey={onCopyKey}
+          account={account}
+          onOpenCCSwitchDialog={onOpenCCSwitchDialog}
+        />
+      }
+      details={{ status: "ready", facts: presentation.detailFacts }}
+      isDetailsExpanded={isExpanded}
+      onDetailsExpandedChange={() => onToggle()}
+      testId={getCopyKeyDialogRuntimeKeyItemTestId(runtimeKey.id)}
+    />
   )
 }
