+++ src/locales/vi/ui.json
@@ -186,7 +186,6 @@
     },
     "collapse": "Thu gọn",
     "copyKey": {
-      "apiKey": "API Key",
       "copied": "Đã sao chép",
       "copy": "Sao chép",
       "copyFailedManual": "Sao chép thất bại, vui lòng sao chép thủ công",
@@ -199,7 +198,6 @@
       "defaultGroup": "Nhóm mặc định",
       "disabled": "Vô hiệu hóa",
       "enabled": "Kích hoạt",
-      "expireTime": "Thời gian hết hạn: {{time}}",
       "exportToCCSwitch": "Xuất sang CC Switch",
       "getFailed": "Lấy dữ liệu thất bại",
       "keyCopied": "Đã sao chép API Key vào khay nhớ tạm",
@@ -208,11 +206,9 @@
       "noKeyFoundAfterCreate": "Tạo thành công nhưng không lấy được API Key, vui lòng thử lại.",
       "noKeys": "Chưa có dữ liệu API Key",
       "noKeysDescription": "Tài khoản này hiện không có API Key khả dụng, bạn có thể tạo nhanh API Key mặc định hoặc tạo tùy chỉnh rồi sao chép.",
-      "remainingQuota": "Hạn mức còn lại",
       "retry": "Thử lại",
       "title": "Danh sách API Key",
       "totalKeys_other": "Tổng cộng {{count}} API Key",
-      "usedQuota": "Hạn mức đã dùng",
       "useInCherry": "Sử dụng trong Cherry Studio"
     },
     "dedupeAccounts": {
